moved functionality of RevengeDialogFactory to RevengeAlertDialogBuilderAdapter

+++ app/src/main/java/com/peternaggschga/gwent/domain/cases/RemoveUnitsUseCase.java
@@ -2,7 +2,6 @@ package com.peternaggschga.gwent.domain.cases;
 
 import android.app.Dialog;
 import android.content.Context;
-import android.content.DialogInterface;
 
 import androidx.annotation.IntRange;
 import androidx.annotation.NonNull;
@@ -72,7 +71,7 @@ public class RemoveUnitsUseCase {
 
     /**
      * Creates a Dialog asking whether the Ability#REVENGE ability should be activated.
-     * The Dialog is created using RevengeDialogFactory.
+     * The Dialog is created using an RevengeAlertDialogBuilderAdapter.
      *
      * @param context      Context of the shown Dialog.
      * @param repository   UnitRepository where the UnitEntity objects are removed and avengers are inserted.
@@ -81,24 +80,23 @@ public class RemoveUnitsUseCase {
      * @param units        Collection of UnitEntity objects that are removed.
      * @param revengeUnits Long representing the number of revenge units.
      * @return A Dialog asking whether the Ability#REVENGE ability should be activated.
-     * @see RevengeDialogFactory#getRevengeDialog(Context, DialogInterface.OnClickListener, DialogInterface.OnClickListener)
-     * @see RevengeDialogFactory#insertAvengers(UnitRepository, int)
+     * @see RevengeAlertDialogBuilderAdapter#insertAvengers(UnitRepository, int)
      */
     @NonNull
     private static Dialog getRevengeDialog(@NonNull Context context, @NonNull UnitRepository repository,
                                            @NonNull CompletableEmitter emitter, @NonNull Collection<UnitEntity> units,
                                            @IntRange(from = 1) int revengeUnits) {
-        return RevengeDialogFactory.getRevengeDialog(context,
-                (dialog, which) -> {
+        return new RevengeAlertDialogBuilderAdapter(context)
+                .setPositiveCallback((dialog, which) -> {
                     // noinspection CheckResult, ResultOfMethodCallIgnored
                     repository.delete(units)
-                            .andThen(RevengeDialogFactory.insertAvengers(repository, revengeUnits))
+                            .andThen(RevengeAlertDialogBuilderAdapter.insertAvengers(repository, revengeUnits))
                             .subscribe(emitter::onComplete);
-                },
-                ((dialog, which) -> {
+                })
+                .setNegativeCallback(((dialog, which) -> {
                     // noinspection CheckResult, ResultOfMethodCallIgnored
                     repository.delete(units).subscribe(emitter::onComplete);
-                })
-        );
+                }))
+                .create();
     }
 }
